remove int usage in protocol classes, update codecs

+++ internal/protocol/list_addall.go
@@ -37,7 +37,7 @@ func ListAddAllEncodeRequest(name *string, valueList []*Data) *ClientMessage {
 	clientMessage.SetMessageType(LIST_ADDALL)
 	clientMessage.IsRetryable = false
 	clientMessage.AppendString(name)
-	clientMessage.AppendInt(len(valueList))
+	clientMessage.AppendInt32(int32(len(valueList)))
 	for _, valueListItem := range valueList {
 		clientMessage.AppendData(valueListItem)
 	}
